With react native web

+++ src/utils/removeEpub.js
@@ -1,10 +1,10 @@
-import { Constants, FileSystem } from "expo"
-import i18n from "./i18n.js"
-import { AsyncStorage } from "react-native"
-import { ActionSheet, Toast } from "native-base"
+import * as FileSystem from 'expo-file-system'
+import Constants from 'expo-constants'
+// import { i18n } from "inline-i18n"
+import { AsyncStorage, Platform } from "react-native"
 
-import { getBooksDir, getSnapshotsDir } from "./toolbox.js"
-import { cancelFetch } from "./zipDownloader.js"
+import { getBooksDir, getSnapshotsDir } from "./toolbox"
+import { cancelFetch } from "./zipDownloader"
 
 const {
   REMOVE_ICON_COLOR,
@@ -38,6 +38,8 @@ const removeEpub = async ({ books, bookId, removeFromBookDownloadQueue, setDownl
 
 export const removeSnapshotsIfANewUpdateRequiresIt = async ({ books, clearAllSpinePageCfis }) => {
 
+  if(Platform.OS === 'web') return
+
   const dateOfMostRecentChangeRequiringPageRecapture = await AsyncStorage.getItem('dateOfMostRecentChangeRequiringPageRecapture')
 
   if(dateOfMostRecentChangeRequiringPageRecapture !== MOST_RECENT_CHANGE_REQUIRING_PAGE_RECAPTURE_DATE) {
@@ -53,6 +55,8 @@ export const removeSnapshotsIfANewUpdateRequiresIt = async ({ books, clearAllSpi
 
 export const clearPageCfiInfoAndSnapshots = async ({ bookId, clearAllSpinePageCfis }) => {
 
+  if(Platform.OS === 'web') return
+
   clearAllSpinePageCfis({ bookId })
   await FileSystem.deleteAsync(`${getSnapshotsDir()}${bookId}`, { idempotent: true })
 
@@ -60,74 +64,80 @@ export const clearPageCfiInfoAndSnapshots = async ({ bookId, clearAllSpinePageCf
 }
 
 export const confirmRemoveEPub = ({ books, bookId, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress, done }) => {
-  ActionSheet.show(
-    {
-      options: [
-        { text: i18n("Remove from device"), icon: "remove-circle", iconColor: REMOVE_ICON_COLOR },
-        { text: i18n("Cancel"), icon: "close" }
-      ],
-      destructiveButtonIndex: 0,
-      cancelButtonIndex: 1,
-      title: i18n(
-        'Are you sure you want to remove "{{book_title}}" from this device?',
-        {
-          book_title: books[bookId].title,
-        }
-      ),
-    },
-    async buttonIndex => {
-      if(buttonIndex == 0) {
-        await removeEpub({ bookId, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress })
-        done && done()
-      }
-    }
-  )
+
+  if(Platform.OS === 'web') return
+
+  // TODO
+  // ActionSheet.show(
+  //   {
+  //     options: [
+  //       { text: i18n("Remove from device"), icon: "remove-circle", iconColor: REMOVE_ICON_COLOR },
+  //       { text: i18n("Cancel"), icon: "close" }
+  //     ],
+  //     destructiveButtonIndex: 0,
+  //     cancelButtonIndex: 1,
+  //     title: i18n("Are you sure you want to remove “{{book_title}}” from this device?", {
+  //         book_title: books[bookId].title,
+  //     }),
+  //   },
+  //   async buttonIndex => {
+  //     if(buttonIndex == 0) {
+  //       await removeEpub({ bookId, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress })
+  //       done && done()
+  //     }
+  //   }
+  // )
 }
 
 export const confirmRemoveAllEPubs = ({ books, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress }) => {
-  ActionSheet.show(
-    {
-      options: [
-        { text: i18n("Remove all books"), icon: "remove-circle", iconColor: REMOVE_ICON_COLOR },
-        { text: i18n("Cancel"), icon: "close" }
-      ],
-      destructiveButtonIndex: 0,
-      cancelButtonIndex: 1,
-      title: i18n("Are you sure you want to remove all books from this device?"),
-    },
-    async buttonIndex => {
-      if(buttonIndex == 0) {
-        await Promise.all(Object.keys(books).map(bookId => (
-          removeEpub({ bookId, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress })
-        )))
-        Toast.show({
-          text: i18n("All books removed."),
-          buttonText: i18n("Okay"),
-          duration: 5000,
-        })
-      }
-    }
-  )
+
+  if(Platform.OS === 'web') return
+
+  // TODO
+  // ActionSheet.show(
+  //   {
+  //     options: [
+  //       { text: i18n("Remove all books"), icon: "remove-circle", iconColor: REMOVE_ICON_COLOR },
+  //       { text: i18n("Cancel"), icon: "close" }
+  //     ],
+  //     destructiveButtonIndex: 0,
+  //     cancelButtonIndex: 1,
+  //     title: i18n("Are you sure you want to remove all books from this device?"),
+  //   },
+  //   async buttonIndex => {
+  //     if(buttonIndex == 0) {
+  //       await Promise.all(Object.keys(books).map(bookId => (
+  //         removeEpub({ bookId, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress })
+  //       )))
+  //       Toast.show({
+  //         text: i18n("All books removed."),
+  //         buttonText: i18n("Okay"),
+  //         duration: 5000,
+  //       })
+  //     }
+  //   }
+  // )
 }
 
 export const confirmRemoveAccountEPubs = ({ books, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress }, callback) => {
-  ActionSheet.show(
-    {
-      options: [
-        { text: i18n("Log out"), icon: "log-out", iconColor: REMOVE_ICON_COLOR },
-        { text: i18n("Cancel"), icon: "close" }
-      ],
-      destructiveButtonIndex: 0,
-      cancelButtonIndex: 1,
-      title: i18n("Are you sure you want to log out and remove all books from this device?"),
-    },
-    async buttonIndex => {
-      if(buttonIndex == 0) {
-        await Promise.all(Object.keys(books).map(bookId => (
-          removeEpub({ books, bookId, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress, removeCover: true })
-        )))
-        callback()
-      }
-    }
-  )
+  // TODO
+  // ActionSheet.show(
+  //   {
+  //     options: [
+  //       { text: i18n("Log out"), icon: "log-out", iconColor: REMOVE_ICON_COLOR },
+  //       { text: i18n("Cancel"), icon: "close" }
+  //     ],
+  //     destructiveButtonIndex: 0,
+  //     cancelButtonIndex: 1,
+  //     title: i18n("Are you sure you want to log out and remove all books from this device?"),
+  //   },
+  //   async buttonIndex => {
+  //     if(buttonIndex == 0) {
+  //       await Promise.all(Object.keys(books).map(bookId => (
+  //         removeEpub({ books, bookId, removeFromBookDownloadQueue, setDownloadStatus, clearTocAndSpines, clearUserDataExceptProgress, removeCover: true })
+  //       )))
+  //       callback()
+  //     }
+  //   }
+  // )
 }
